chore: 프로젝트 기본 파일 그룹화
