+++ WeatherForecast/WeatherForecast/Configuration/Assets.xcassets/AppIcon.appiconset/Contents.json
