+++ WeatherForecast/WeatherForecast/Configuration/Assets.xcassets/Contents.json
