+++ WeatherForecast/WeatherForecast/Configuration/Info.plist
